rearrange tests to be discovered on dj18
