+++ testproject/testproject/settings.py
@@ -125,9 +125,10 @@ INSTALLED_APPS = (
     # Uncomment the next line to enable admin documentation:
     # 'django.contrib.admindocs',
     'phonenumber_field',
-    'phonenumber_field.tests',
     'testapp',
 )
+
+MIGRATION_MODULES = dict((app, '%s.fake_migrations' % app) for app in INSTALLED_APPS)
 
 # A sample logging configuration. The only tangible logging
 # performed by this configuration is to send an email to
